allow to select attached pci devices and use mac address pool

+++ main.py
@@ -3,7 +3,7 @@ from contextlib import ExitStack
 import sys
 
 from qemu import QemuVm
-from virtual_machines import virtual_machines
+from virtual_machines import use_mac_address_from_pool, virtual_machines
 from hosts import get_conn, is_reachable, wake
 from repl import Repl
 from net import find_next_free_port
@@ -15,6 +15,7 @@ if __name__ == "__main__":
         description="spawn and enter into bugbakery testing VMs",
     )
     parser.add_argument("vm", choices=virtual_machines.keys())
+    parser.add_argument("-d", "--devices")
     parser.add_argument("-v", "--verbose", action="store_true")
     parser.add_argument("--vnc", action="store_true")
     parser.add_argument("--no-rdp", action="store_true")
@@ -32,29 +33,41 @@ if __name__ == "__main__":
     wake(vm_host)
 
     vm_config = virtual_machine_spec["vm_config"]
-    vm = QemuVm(
-        **vm_config,
-        memory="6G",
-        cores=4,
-        ephemeral=not args.edit_base_image,
-        vnc_display=0 if args.vnc else None,
-    )
+
+    if args.devices:
+        devices = sum([virtual_machine_spec["available_devices"][dev] for dev in args.devices.split(',')], [])
+    else:
+        devices = []
 
     host_conn = get_conn(vm_host)
 
     try:
-        cmd = vm.build_cmd()
-        if verbose:
-            print(f"Running command: {cmd}")
-        vm_cmd_result = host_conn.sudo(
-            cmd,
-            password=host_conn.connect_kwargs["password"],
-            asynchronous=True,
-            out_stream=sys.stdout if verbose else None,
-            err_stream=sys.stderr if verbose else None,
-            pty=True,  # needed, so the command is killed properly when exiting
-        )
         with ExitStack() as stack:
+            mac, ip = stack.enter_context(use_mac_address_from_pool(virtual_machine_spec["host"]))
+            virtual_machine_spec["wireguard"] = ip
+
+            vm = QemuVm(
+                **vm_config,
+                memory="6G",
+                cores=4,
+                ephemeral=not args.edit_base_image,
+                vnc_display=0 if args.vnc else None,
+                pci_devices=devices,
+                mac_address=mac,
+            )
+
+            cmd = vm.build_cmd()
+            if verbose:
+                print(f"Running command: {cmd}")
+
+            vm_cmd_result = host_conn.sudo(
+                cmd,
+                password=host_conn.connect_kwargs["password"],
+                asynchronous=True,
+                out_stream=sys.stdout if verbose else None,
+                err_stream=sys.stderr if verbose else None,
+                pty=True,  # needed, so the command is killed properly when exiting
+            )
             if not args.skip_boot_check:
                 print("-> Waiting for VM to boot...")
                 while not is_reachable(virtual_machine_spec):
@@ -100,4 +113,5 @@ if __name__ == "__main__":
 
     finally:
         print("-> Stopping vm...")
-        vm_cmd_result.runner.kill()
+        if vm_cmd_result:
+            vm_cmd_result.runner.kill()
